Add original veth.c code
Add source code for the original "veth.c" module from the Linux kernel
tree. This was pulled from the Linus -rc tree at the time of this
commit.

For reference purposes: 'git log' indicates the last change on this file
was commit 5517750f058edd111bcabe5e116056cc63b1f39c .

The minimal Makefile included in this commit is sufficient to recompile
the module as-is out of tree.